extract combinations logic from file
